requirements.txt
pillow==11.0.0
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.10.2
python-multipart==0.0.17
sqlalchemy==2.0.36
bcrypt==4.2.1
